requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
pydantic-settings>=2.4
pytest>=8.0
httpx>=0.27
